Fix Clang warnings

+++ src/core/runtime/include/cassian/runtime/image_properties.hpp
@@ -45,7 +45,7 @@ struct ImageDimensions {
    *
    * @param width
    */
-  constexpr ImageDimensions(const uint32_t width) : width(width) {}
+  constexpr ImageDimensions(const uint64_t width) : width(width) {}
 
   /**
    * Construct a new Image Dimensions object
@@ -53,7 +53,7 @@ struct ImageDimensions {
    * @param width
    * @param height
    */
-  constexpr ImageDimensions(const uint32_t width, const uint32_t height)
+  constexpr ImageDimensions(const uint64_t width, const uint32_t height)
       : width(width), height(height){};
 
   /**
@@ -63,7 +63,7 @@ struct ImageDimensions {
    * @param height
    * @param depth
    */
-  constexpr ImageDimensions(const uint32_t width, const uint32_t height,
+  constexpr ImageDimensions(const uint64_t width, const uint32_t height,
                             const uint32_t depth)
       : width(width), height(height), depth(depth) {}
